finalize response

+++ api/src/main/kotlin/li/doerf/microstore/api/rest/dto/CreateOrderResponse.kt
@@ -5,7 +5,7 @@ import java.math.BigDecimal
 
 data class CreateOrderResponse(
         val orderId: String,
-        val orderNumber: Int,
+        val orderNumber: Long,
         val amount: BigDecimal,
         val orderStatus: OrderEndState,
         val errorText: String
